# OpenAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Realtime Chat",
    "description": "API de chat em tempo real.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:8000"
    }
  ],
  "paths": {
    "/auth/refresh": {
      "get": {
        "description": "Atualiza uma autenticação com o usuário, invalidando o Token de e gerando um novo",
        "parameters": [
          {
            "name": "Authoriarion",
            "in": "header",
            "required": true,
            "schema": {"type": "string"},
            "description": "Token de autenticação no formato Bearer"
          }
        ],
        "responses": {
          "200": {
            "description": "Novo token gerado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1dWlkIjoiMWE0ZGUzODAtOWZhOS00ZmU5LWJjZGItZDQ1Y2NhNDdmNzk0IiwiZGVzdCI6IjhkODdiMTliLTViNzYtNDI4MS1iZWQ0LTU5MGY3NzUyMjM0OCIsImF0IjoiMjAyNC0wNS0zMSAyMTozOToxOS40MzM2NTEifQ.DdRqNhi_qI7fcYVmXhFsLd8cZ5bfL2hzCinHEqyz_I0"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Pode ocorrer quando:\n\tO Token não pôde ser decodificado, sendo considerado inválido;\nO uuid de usuário contido no Token não é registrado no banco de dados\n;Não existe um protocolo equivalente com o fornecido pelo Token, com os protocolos contidos no banco de dados.",
            "content": {"text/plain": {"example": "Unauthorized"}}
          },
          "419": {
            "description": "O token fornecido já foi expirado! Necessário fazer autenticação novamente.",
            "content": {"text/plain": {"example": "Connection will expire! authenticate again."}}
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "description": "Registrar um usuário",
        "responses": {
          "201": {"description": "Registro realizado com sucesso"},
          "409": {
            "description": "E-mail fornecido já foi registrado",
            "content": {"text/plain": {"example": "E-mail already registered!"}}
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {"type": "string", "example": "exemple@email.com"},
                  "password": {"type": "string", "example": "12345"},
                  "username": {"type": "string", "example": "Exemple"}
                },
                "required": ["email", "password", "username"]
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "Realiza autenticação, retornando um Token JWT",
        "responses": {
          "200": {
            "description": "Autenticação realizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1dWlkIjoiMWE0ZGUzODAtOWZhOS00ZmU5LWJjZGItZDQ1Y2NhNDdmNzk0IiwiZGVzdCI6IjhkODdiMTliLTViNzYtNDI4MS1iZWQ0LTU5MGY3NzUyMjM0OCIsImF0IjoiMjAyNC0wNS0zMSAyMTozOToxOS40MzM2NTEifQ.DdRqNhi_qI7fcYVmXhFsLd8cZ5bfL2hzCinHEqyz_I0"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais de autenticação incorreta",
            "content": {"text/plain": {"example": "Unauthorized!"}}
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {"type": "string", "example": "exemple@email.com"},
                  "password": {"type": "string", "example": "12345"}
                },
                "required": ["email", "password"]
              }
            }
          }
        }
      }
    }
  }
}